requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.3
django-bootstrap3==23.6
django-bootstrap4==24.1
django-crispy-forms==1.12.0
django-etc==1.4.0
django-heroku==0.3.1
django-hitcount==1.3.5
django-resized==1.0.2
django-taggit==5.0.1
django-tinymce==4.0.0
gunicorn==21.2.0
idna==3.6
ndg-httpsclient==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.6.0
pycparser==2.22
PyNaCl==1.5.0
pyOpenSSL==24.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
